requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
Werkzeug==3.0.1
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
langchain==0.1.0
langchain-groq==0.1.0
groq==0.4.1
reportlab==4.0.7
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
